+++ Exp_femto_1d/app/main.cpp
@@ -74,6 +74,7 @@ int main(int argc, char **argv) {
       std::cout << "build-cf"
                 << " stored_slices=" << statistics.stored_slices
                 << " skipped_zero_mixed_event_groups=" << statistics.skipped_zero_mixed_event_groups
+                << " skipped_zero_mixed_event_slices=" << statistics.skipped_zero_mixed_event_slices
                 << " skipped_zero_same_event_slices=" << statistics.skipped_zero_same_event_slices << "\n";
       return 0;
     }