+++ Exp_femto_1d/scripts/cmake.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+cmake -S /Users/allenzhou/Research_software/Code_base/Exp_femto_1d -B /Users/allenzhou/Research_software/Code_base/Exp_femto_1d/build
+
+cmake --build /Users/allenzhou/Research_software/Code_base/Exp_femto_1d/build -j 8